pyproject.toml
[tool.poetry]
name = "learn-django"
version = "0.1.0"
description = "learn django"
authors = ["aprilahijriyan <hijriyan23@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
Django = "^3.2"
djangorestframework = "^3.12.4"
bs4 = "^0.0.1"
django-filter = "^2.4.0"
cryptography = "^3.4.7"
djangorestframework-simplejwt = "^4.6.0"
django-extensions = {version = "^3.1.3", extras = ["full"]}
drf-yasg = "^1.20.0"

[tool.poetry.dev-dependencies]
ipython = "^7.22.0"
pre-commit = "^2.12.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
